The Front-End Of EAFT
I created six front-ends for the project. In one page(Main.html) I had to use bootstrap for the Ad-Block.

There are two CSS files. One is bootstrap.css which is used in one page. And style.css is used in all pages. There are some images and logos that I used in pages.